Prueba_Final/requirements.txt
Flask==2.2.5
Werkzeug==2.2.3
Pillow==10.0.0
